requirements.txt
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.3
cloudinary==1.26.0
cryptography==3.4.6
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==2.2.8
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycosat==0.6.3
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.58.0
urllib3==1.25.11
whitenoise==5.2.0
